requirements-dev.txt
PyHamcrest ~= 2.0.3
hypothesis ~= 6.43.2
black ~= 22.6.0
